🚚 move june analytics to folder and update paths
