+++ frontend/app/brains-management/[brainId]/components/BrainManagementTabs/components/KnowledgeTab/DocumentItem/DocumentData.tsx
@@ -3,7 +3,7 @@ import { useEffect, useState } from "react";
 
 import { useAxios } from "@/lib/hooks";
 
-import { useEventTracking } from "@/services/analytics/useEventTracking";
+import { useEventTracking } from "@/services/analytics/june/useEventTracking";
 import { useTranslation } from "react-i18next";
 import { useSupabase } from "../../../../../../../../lib/context/SupabaseProvider";
