+++ frontend/lib/components/NavBar/components/NavItems/components/LanguageDropDown/hooks/useLanguageHook.ts
@@ -1,7 +1,7 @@
 import { useEffect, useState } from "react";
 import { useTranslation } from "react-i18next";
 
-import { useEventTracking } from "@/services/analytics/useEventTracking";
+import { useEventTracking } from "@/services/analytics/june/useEventTracking";
 
 export type Language = {
   id: string;
